portfolio-advisor/pyproject.toml
[tool.poetry]
name = "portfolio-advisor"
version = "0.1.0"
description = "AI options portfolio advisor for TastyTrade & ThinkOrSwim"
authors = ["Trader"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
anthropic = "^0.34.0"
pandas = "^2.2.0"
numpy = "^1.26.0"
python-multipart = "^0.0.9"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
yfinance = "^0.2.40"
aiofiles = "^23.2.1"
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
